@@ -1,4 +1,5 @@
 #include "libuv_net/thread_pool.hpp"
+#include <spdlog/spdlog.h>
 
 namespace libuv_net {
 
@@ -9,14 +10,12 @@ ThreadPool::ThreadPool(size_t num_threads) {
                 std::function<void()> task;
                 {
                     std::unique_lock<std::mutex> lock(queue_mutex_);
-                    condition_.wait(lock, [this] { 
-                        return stop_ || !tasks_.empty(); 
+                    condition_.wait(lock, [this] {
+                        return stop_ || !tasks_.empty();
                     });
-                    
                     if (stop_ && tasks_.empty()) {
                         return;
                     }
-                    
                     task = std::move(tasks_.front());
                     tasks_.pop();
                 }
@@ -33,10 +32,20 @@ ThreadPool::~ThreadPool() {
     }
     condition_.notify_all();
     for (auto& thread : threads_) {
-        if (thread.joinable()) {
-            thread.join();
-        }
+        thread.join();
     }
+}
+
+void ThreadPool::enqueue(std::function<void()> task) {
+    {
+        std::unique_lock<std::mutex> lock(queue_mutex_);
+        if (stop_) {
+            spdlog::error("ThreadPool is stopped, cannot enqueue task");
+            return;
+        }
+        tasks_.emplace(std::move(task));
+    }
+    condition_.notify_one();
 }
 
 } // namespace libuv_net 
